==> notebooks/1 - sample generating.ipynb <==
# Sample generating
In dit notebook gaan wij een sample genereren met een bepaald aantal eisen.

## Eisen sample

De sample heeft een aantal eisen waar aan voldaan moeten worden:
- De sample bevat data van het jaar 2020.
- De sample bevat de features `date`, `serial_number`, `model`, `failure`, `smart_5_raw`, `smart_9_raw`, `smart_187_raw`, `smart_188_raw`, `smart_194_raw`, `smart_197_raw`, en `smart_198_raw`.

## Libraries importeren

import glob
import pandas as pd

## Data importeren
De gebruikte data zijn csv files van alle dagen uit 2020. Dit zijn dus in totaal 366 csv files.

Gezien er ontzettend veel data gebruikt wordt, hebben wij deze data opgesplitst en per maanden ingeladen (1 keer voor de maanden januari - mei, 1 keer voor de maanden juni - november, en als laatst 1 keer voor de maand december).

path = '../data/raw'
all_files = glob.glob(path + "/*.csv")

df = pd.concat((pd.read_csv(f) for f in all_files))

df.shape

## Onderscheid maken SSD's
Gezien er zowel SSD's als HDD's in de dataset zitten, moet hier onderscheid tussen worden gemaakt. Wij gaan namelijk alleen voorspellen over data van HDD's.

Als eerste wordt er een extra kolom `ssd` toegevoegd aan de dataset, en staat deze standaard op `False`.

df.insert(2, 'ssd', False, True)

Vervolgens wordt er gekeken naar `smart_233_raw`. Dit is een SMART value die alleen voor komt bij SSD's. Wanneer deze dus leeg is, is die entry geen SSD.
Bij alle entries waarbij `smart_233_raw` niet leeg is, wordt binnen de kolom SSD `True` gezet. Vervolgens worden al deze entries uit de dataframe verwijderd.

df.loc[df['smart_233_raw'].notna(), 'ssd'] = True
df.drop(df[df['ssd'] == True].index, inplace=True)

df.shape

df.head()

## Columns droppen
Bepaalde columns worden niet meegenomen in onze voorspellingen. Deze worden uit de dataframe verwijderd.

df.drop(df.columns.difference(['date', 'serial_number', 'model', 'failure', 'smart_5_raw', 'smart_9_raw', 'smart_187_raw', 'smart_188_raw', 'smart_194_raw', 'smart_197_raw', 'smart_198_raw']), 1, inplace=True)

df.head()

## Data exporteren

Uiteindelijk exporteren wij de dataframe naar een csv file.

df.to_csv('..\data\processed\sample.csv')

## Data mergen

Gezien wij 3 verschillende csv files met verschillende maanden gegenereerd hebben, hebben wij de csv's samengevoegd met onderstaand stukje code.

#df = pd.concat([first_df, second_df], ignore_index=True)
