==> notebooks/2 - sample analysing.ipynb <==
# Data analyse en processing
In dit notebook gaan wij kijken naar de sample set die gegenereerd is in een eerder notebook (`1 - sample`) en hoe wij hiermee verder kunnen gaan.

## Libraries importeren

import pandas as pd
import seaborn as sn
import matplotlib.pyplot as plt

## Data importeren

De data die gebruikt wordt is de data die gegenereerd is in het notebook `1 - sample generating`.

df = pd.read_csv('../data/processed/sample.csv')

## Data valideren
Om te kijken of de sample set correct geprocessed is, gaan wij deze eerst valideren.

Ten eerste kijken wij naar het begin van de dataframe, middels de functie `head()`, en naar het einde van de dataframe, middels de functie `tail()`. Hiermee valideren wij dat de data begint op 1 januari 2020 (2020-01-01) en eindigt op 31 december 2020 (2020-12-31).

df.head()

df.tail()

Ook kijken wij naar het aantal unieke `date` entries. Dit moet gelijk staan aan 366, gezien 2020 een schrikkeljaar was en daardoor 366 dagen had.

df['date'].drop_duplicates().count()

Vervolgens kijken wij nog naar het aantal entries in de sample set.

df.shape

## NaN waardes
In de sample set zitten gegarandeerd NaN waardes. NaN staat voor 'Not a Number'.
In onze sample set willen wij zo min mogelijk NaN waardes hebben, omdat wij met NaN waardes onze predicties niet kunnen uitvoeren.

Ten eerste gaan wij kijken hoeveel NaN waardes elke feature heeft.

df.isnull().sum()

df.isnull().sum().sum()

Dit komt uiteindelijk uit op een totaal van `20061528` NaN waardes in de hele sample set.

### Procent van missing NaN waardes per feature weergeven
Om een beter beeld te krijgen van de NaN waardes, maken wij een overzicht met het procent van NaN waardes per feature, ten opzichten van het totaal aantal entries van die feature.

missing_nan = df.isnull().sum() * 100 / len(df)
missing_nan_df = pd.DataFrame({'% NaN': missing_nan})
missing_nan_df.head(15)

Uit bovenstaand overzicht kunnen wij concluderen dat de meeste NaN waardes binnen de features `smart_187_raw` en `smart_188_raw` zitten.

### NaN waardes per model
Om een nog beter beeld te krijgen over de NaN waardes, gaan wij een overzicht maken.

Dit overzicht geeft per model, het percentage NaN waardes ten opzichten van het totale aantal entries, per feature weer.

df_nan = df.isnull().groupby(df.model).mean().round(4).mul(100)

df_nan = df_nan.drop(['date', 'serial_number', 'model', 'failure'], axis = 1)
df_nan.head(51)

Uit bovenstaand overzicht kunnen wij een aantal dingen concluderen:

- Elk model heeft, of een NaN percentage van onder de 1% (met uitzondering van `WDC WUH721414ALE6L4`, die op 1.13% zit), of een NaN percentage van 100% bij `smart_187_raw` en `smart_188_raw` (met uitzondering van `TOSHIBA MG07ACA14TEY` en `TOSHIBA MG08ACA16TEY`)
- De modellen die 100% NaN hebben bij `smart_187_raw`, hebben ook 100% NaN bij `smart_188_raw`
- Het model `DELLBOSS VD` is geen schijf

Ten eerste halen wij het model `DELLBOSS VD` uit de dataframe, gezien dit geen schijf is.

df = df.drop(df[df.model == 'DELLBOSS VD'].index)

df[df.model == 'DELLBOSS VD'].sum()

Ook halen wij `DELLBOSS VD` uit de `df_nan` set.

df_nan = df_nan.drop('DELLBOSS VD')

Vervolgens is het voor ons handig om te zien, hoeveel entries elk model heeft ten opzichte van het totaal aantal entries.

Eerst voegen wij een kolom `amount_entries` toe aan de dataframe.

df_nan.insert(0, "amount_entries", float)

df_nan.head()

Vervolgens gaan wij elk model in de dataframe `df_nan` af. 

Per model kijken wij hoeveel entries dit model heeft in de dataframe `df`. Dit delen wij vervolgens door het totaal aantal totale entries, en doen dit getal keer 100. Hiermee krijgen wij een percentage.

Dit percentage wordt dan opgeslagen in de kolom `amount_entries` bij het desbetreffende model.

for i in df_nan.index:
    df_nan.at[i, 'amount_entries'] = (((len(df[df['model'] == i])) / df.shape[0]) * 100)

df_nan.head(50)

Om te valideren dat bovenstaande gegevens kloppen, tellen wij alle waardes van `amount_entries` bij elkaar op. De som hiervan moet uitkomen op 100.

df_nan['amount_entries'].sum()

Als laatste halen wij de `TOSHIBA MG07ACA14TEY` en `TOSHIBA MG08ACA16TEY` modellen eruit. Deze modellen hebben bij `smart_197_raw` beide een missing NaN percentage van 100%.

Gezien deze modellen niet al te veel entries in de dataset hebben, halen we deze uit de dataframes.

df = df.drop(df[df.model == 'TOSHIBA MG07ACA14TEY'].index)
df = df.drop(df[df.model == 'TOSHIBA MG08ACA16TEY'].index)

df_nan = df_nan.drop('TOSHIBA MG07ACA14TEY')
df_nan = df_nan.drop('TOSHIBA MG08ACA16TEY')

df_nan.head(50)

## Data opsplitsen
Om de data bruikbaar te maken, splitsen wij deze op in 2 verschillende sets.

Zoals in bovenstaand overzicht te zien is, zijn er veel modellen die alleen `smart_187_raw` en `smart_188_raw` missen. Deze modellen zetten wij in een aparte dataframe.

Ten eerste gaan we kijken welke schijven een NaN percentage van 100% hebben bij `smart_187_raw`. De schijven die een percentage van 100% NaN hebben bij `smart_187_raw`, hebben ook een 100% NaN percentage bij `smart_188_raw`. 

Om dit te doen gebruiken wij de dataframe `df_nan`, die eerder gebruikt is. Hieruit halen wij de modellen die een NaN percentage van 100% hebben.

De entries van deze modellen voegen wij vervolgens toe aan een dataframe `df_2`, en maken nog een dataframe `df_1` aan met alle modellen die niet een NaN percentage van 100% hebben.

df_nan_models = df_nan[df_nan['smart_187_raw'] == 100].index

df_2 = df[df['model'].isin(df_nan_models)]
df_2.shape[0]

df_1 = df[~df['model'].isin(df_nan_models)]
df_1.shape[0]

df.shape[0]

Het totaal aantal entries van de originele dataframe `df` was `31925780`.

Wanneer wij de entries van `df_1` (`22008389`) en `df_2` (`9917391`) optellen, komen wij uit op het volgende:

if (df_1.shape[0] + df_2.shape[0] == df.shape[0]):
    print(True)

else:
    print(False)

Dit betekent dat de som van `df_1` en `df_2` gelijk staat aan `df`.

Hieruit kunnen wij concluderen dat bovenstaande code de modellen correct heeft opgesplitst.

## Data exporteren

Uiteindelijk exporteren wij de twee verschillende dataframes naar csv files.

df_1.to_csv('..\\data\\processed\\sample_without_100nan.csv', index = False)

df_2.to_csv('..\\data\\processed\\sample_with_100nan.csv', index = False)
