==> notebooks/6 - support vector machine.ipynb <==
# Support Vector Machine
In deze notebook wordt er een support vector machine opgesteld en getrained voor het voorspellen van failure bij harde schijven.

Stap 1: Libraries importeren

import pandas as pd
import numpy as np
import pickle
import os

from sklearn.svm import SVC
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score, plot_confusion_matrix


# Support Vector Machine

Een Support Vector Machine (SVM) is een model dat aan de hand van verschillende kenmerken objecten kan
classificeren. Het model maakt gebruikt van een lijn (in 2D) of een hypervlak (Hyperplane) om zo punten in
een vectorruimte te scheiden in klassen. Een SVM-algoritme maakt de scheiding op basis van de marge zo
groot mogelijk en gelijkmatig te maken tussen de meest nabijgelegen kenmerken van een klasse. De naam
komt van de meest nabijgelegen kenmerken die ook wel support vectors worden genoemd.

De verschillende SMART waardes 5, 9, 187, 188, 194, 197 en 198.ie er worden ingevoerd om het model te maken zijn de waardes: 

## Data importeren

De gebruikte data is een dataset van de maand januari.

dataset = pd.read_csv('../data/processed/sample_without_100nan.csv')

# Data cleaning
In de dataset zitten nog een aantal kolommen die niet gebruikt hoeven te worden om het model te trainen en geen invloed hebben op de voorspelling van de failure van een hdd. Deze worden daarom ook eruit gehaald.

del dataset['date']
del dataset['serial_number']
del dataset['model']

dataset.shape

Er zitten nog enkele lege waardes in de dataset. Na een analyse blijkt dat het om ongeveer 1% van de rijen gaat. Deze worden daarom dus verwijderd, hierna blijven er nog een groot aantal rijen over om te trainen.

assert isinstance(dataset, pd.DataFrame), "df needs to be a pd.DataFrame"
dataset.dropna(inplace=True)
indices_to_keep = ~dataset.isin([np.nan, np.inf, -np.inf]).any(1)
dataset[indices_to_keep].astype(np.float64)

De failure wordt op de y as geplaatst en de rest van de SMART waardes op de X as.

X = dataset.drop(['failure'], axis=1).values
y = dataset['failure'].values

X.shape, y.shape

# Trainings- en testdata
De data wordt hier opgesplits in een trainingsset en een testset in een verhouding van 80/20.

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=0.4)

X_train.shape, y_train.shape

X_test.shape, y_test.shape

# Trainen van het model
Hier wordt het SVM model getrained. DIt wordt gedaan middels het Sklearn pakket. 

svm = SVC(kernel='linear')
svm.fit(X_train, y_train)
